+++ dbInit.js
@@ -1,18 +1,22 @@
 const initDB = async (reset_db) => {
     try {
-        await require('./database/dbInit/full.init')(reset_db);
-        await require('./database/dbInit/server.init')(reset_db);
+        inits = [ require('./database/dbInit/userdata.init')(reset_db),
+                require('./database/dbInit/serverdata.init')(reset_db)];
+        await Promise.all(inits);
     } catch (e) {
         console.log(e)
     };
 };
 
-let argv = require('minimist')(process.argv.slice(2));
-if(argv.delete){
-    if(argv.delete.toLowerCase()=='true')
-    console.log('All data will be deleted')
-    initDB(true);
-} else {
-    initDB(false);
-}
 
+
+(async function main () {
+    let argv = require('minimist')(process.argv.slice(2));
+    if(argv.delete){
+        if(argv.delete.toLowerCase()=='true')
+        console.log('All data will be deleted')
+        await initDB(true);
+    } else {
+        await initDB(false);
+    }
+})();
